[Context] Verify logging worker before storing run
